Add Service to count step, dependancy injection to get the number of step in heart.
