+++ package.json
@@ -36,6 +36,7 @@
     "cordova-plugin-device-motion": "^2.0.1",
     "cordova-plugin-geolocation": "^4.0.2",
     "cordova-plugin-gyroscope": "^0.1.4",
+    "cordova-plugin-pip": "0.0.2",
     "cordova-sqlite-storage": "^4.0.0",
     "core-js": "^2.5.4",
     "d3": "^5.15.0",
@@ -97,7 +98,8 @@
       "cordova-plugin-ionic-keyboard": {},
       "cordova-sqlite-storage": {},
       "cordova-plugin-geolocation": {},
-      "cordova-plugin-background-mode": {}
+      "cordova-plugin-background-mode": {},
+      "cordova-plugin-pip": {}
     },
     "platforms": [
       "android"
