+++ src/app/heart/heart.page.html
@@ -13,7 +13,7 @@
       <h2 class="title bas ">
        PAS
       </h2>
-      <h2 class="title haut  text_number">{{ step }}</h2>
+      <h2 class="title haut  text_number" >{{ step }}</h2>
       <section class="stage">
         <figure class="ball"><span class="shadow"></span></figure>
       </section>
